// shared/src/types/tmdb.ts
export type WatchProvider = {
  provider_id: number
  provider_name: string
  logo_path: string
  display_priority: number
}

export type WatchProvidersResult = {
  results: {
    ES?: {
      flatrate?: WatchProvider[]
      rent?: WatchProvider[]
      buy?: WatchProvider[]
    }
  }
}